change unsigned in to size_t
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@1312 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/moses/src/LexicalReordering.h
@@ -36,7 +36,7 @@ class LexicalReordering : public ScoreProducer {
   virtual ~LexicalReordering();
  public: //interface
   //inherited
-  virtual unsigned int GetNumScoreComponents() const {
+  virtual size_t GetNumScoreComponents() const {
     return m_NumScoreComponents; 
   };
   
@@ -58,7 +58,7 @@ class LexicalReordering : public ScoreProducer {
   Phrase auxGetContext(const Hypothesis* hypothesis) const;
  private:
   LexicalReorderingTable* m_Table;
-  int m_NumScoreComponents;
+  size_t m_NumScoreComponents;
   std::vector< Direction > m_Direction;
   std::vector< Condition > m_Condition;
   bool m_OneScorePerDirection;
